requirements.txt
numpy>=2.2.4
tk>=0.1.0
pillow>=10.4.0
scikit-learn>=1.6.1
